refactor: 💡 cli with source arugment
We're going to have CSVs from many different places. This sets up up to
pass in the source, in lowercase, of the csv we're passing in to
execute, and only need to add a new export in the formatters folder to
extend to new CSV normalization methods.

BREAKING CHANGE: 🧨 command now expects two arguments and accepts 3 - [csvfile] [csvsource]
[flag?]